Upgrade spring-boot-starter-parent 2.1.8.RELEASE to 2.2.1.RELEASE (#8743)
* Upgrade spring-boot-starter-parent 2.1.8.RELEASE to 2.2.0.RELEASE

* Upgrade spring-boot-starter-parent 2.2.0.RELEASE to 2.2.1.RELEASE

* attempt to fix platform it
