+++ molgenis-app/src/main/java/org/molgenis/app/WebAppSecurityConfig.java
@@ -9,13 +9,13 @@ import org.molgenis.data.DataService;
 import org.molgenis.data.security.DataserviceRoleHierarchy;
 import org.molgenis.data.support.DataServiceImpl;
 import org.molgenis.security.MolgenisWebAppSecurityConfig;
+import org.molgenis.security.acl.AclConfig;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import org.springframework.security.access.AccessDecisionVoter;
 import org.springframework.security.access.hierarchicalroles.RoleHierarchy;
 import org.springframework.security.access.vote.AffirmativeBased;
-import org.springframework.security.acls.jdbc.AclConfig;
 import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
