+++ molgenis-dataexplorer/pom.xml
@@ -68,7 +68,7 @@
     </dependency>
     <dependency>
       <groupId>org.glassfish</groupId>
-      <artifactId>javax.el</artifactId>
+      <artifactId>jakarta.el</artifactId>
     </dependency>
     <dependency>
       <groupId>org.apache.commons</groupId>
