+++ pom.xml
@@ -7,7 +7,7 @@
   <parent>
     <groupId>org.springframework.boot</groupId>
     <artifactId>spring-boot-starter-parent</artifactId>
-    <version>2.1.8.RELEASE</version>
+    <version>2.2.1.RELEASE</version>
   </parent>
   <packaging>pom</packaging>
 
@@ -81,8 +81,6 @@
     <!-- override dependency versions managed by parent pom -->
     <elasticsearch.version>5.5.3</elasticsearch.version>
     <guava.version>25.0-jre</guava.version>
-    <mockito.version>3.0.0</mockito.version>
-    <junit-jupiter.version>5.5.2</junit-jupiter.version>
     <!-- parent version 0.35 depends on org.apache.commons:commons-compress:jar:1.9, 1.18 is required by org.apache.poi:poi:jar:4.0.1 -->
     <webjars-locator-core.version>0.36</webjars-locator-core.version>
     <!-- Java 11 support -->
